test-deployments/prod-sovereign-agent/app/agent.py
import os
import re
from typing import List, Any
from google.adk.agents import Agent
from google.adk.apps import App
from google.adk.agents.context_cache_config import ContextCacheConfig
from pydantic import PrivateAttr

# v1.4.7 Sovereign Alignment: Native ADK Implementation
# Hydrated via Agent Starter Pack & Hardened for Production

class SuperAgent(Agent):
    """
    My Super Agent: Native ADK Implementation.
    Hardened for Production with Guardrails and Context Caching.
    """
    name: str = "my_super_agent"
    description: str = "High-fidelity AI agent for solving complex laboratory tasks."
    
    # Internal state
    _forbidden_patterns: List[str] = PrivateAttr(default_factory=lambda: [
        r"(?i)password", r"(?i)secret", r"(?i)social security", r"\d{3}-\d{2}-\d{4}"
    ])

    def model_post_init(self, __context: Any) -> None:
        # FinOps: Enable Context Caching (3600s TTL)
        try:
            from google.adk.agents.context_cache_config import ContextCacheConfig
            self.context_cache_config = ContextCacheConfig(ttl_seconds=3600)
        except Exception:
            pass
        super().model_post_init(__context)

    def _sanitize_input(self, text: str) -> str:
        """Security: Guard against prompt injection and PII leakage."""
        if not text: return ""
        if "ignore all previous instructions" in text.lower():
            return "REJECTED: Security bypass blocked."
        
        sanitized = text
        for pattern in self._forbidden_patterns:
            sanitized = re.sub(pattern, "[REDACTED]", sanitized)
        return sanitized

    async def _run_async_impl(self, ctx):
        """Native ADK Async Execution."""
        # Significant logic would go here
        from google.adk.events import Event
        
        user_input = ctx.session.last_user_message.parts[0].text
        safe_input = self._sanitize_input(user_input)
        
        if "REJECTED" in safe_input:
             yield Event(author=self.name, content="Safety block triggered.")
             return

        response = f"Sovereign Solution: Processed '{safe_input}' with ADK Native Reasoning."
        yield Event(author=self.name, content=response)

# Create the root agent
root_agent = SuperAgent()

# Wrap in ADK App
app = App(root_agent=root_agent, name="my_super_agent")
